# backend/openapi/v2.yaml
openapi: 3.0.3
info:
  title: Tem de Tudo API
  version: 2.0.0
  description: >
    Especificacao OpenAPI v2 com governanca de privacidade (LGPD),
    conciliacao de billing e idempotencia estrita em producao.
servers:
  - url: /api
...
paths:
...
  /privacy/export/{privacyRequestId}/download:
    get:
      tags: [Privacy]
      summary: Download do pacote de dados pessoais
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: privacyRequestId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Arquivo JSON de exportacao
        "404":
          description: Solicitacao nao encontrada
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum